Ensure that headers initially marked CXX can be later marked mixed.
When a header was marked CXX, it wouldn't be parsed for CPP_IDENTIFIERS
and then wouldn't be opened or read past that point. This meant that
it could never be marked as a mixed mode header, which was a bug.

Use a new map to determine which headers are mixed mode and scan all
headers (not just C ones) for CPP_IDENTIFIERS. The mixed mode
language will be determined at the call site of
_psq_get_absolute_path_to_header_file_language